feat: connecting ui components
