+++ src/page/template/panel/guest-options.htm
@@ -7,7 +7,7 @@
     </div>
     <div class="panel__content" data-bind="fadingscrollbar">
       <div class="guest-options__content">
-        <guest-mode-toggle params="isChecked: isGuestEnabled(), onToggle: toggleAccessState, isDisabled: requestOngoing"></guest-mode-toggle>
+        <guest-mode-toggle params="isChecked: isGuestEnabled(), onToggle: toggleServiceAccessState, isDisabled: requestOngoing"></guest-mode-toggle>
       </div>
       <!-- ko if: showLinkOptions() -->
         <div class="guest-options__info panel__info-text">
